metrique 1, 2, 3, 4, 5, 9

+++ tp1/parsingUI/src/packageModels/Metricable.java
@@ -5,7 +5,7 @@ import java.util.*;
 public interface Metricable {
 	
 	//ANA
-	int getAverageMethodArgumentCount();
+	float getAverageMethodArgumentCount();
 		
 	//NOM
 	int getMethodCount();
